static analysis updates

+++ osCommerce/OM/Custom/Site/Website/Application/Account/Action/ChangePassword.php
@@ -18,7 +18,7 @@ class ChangePassword
     public static function execute(ApplicationAbstract $application)
     {
         if (!isset($_SESSION[OSCOM::getSite()]['Account'])) {
-            $req_actions = http_build_query(array_slice($_GET, array_search('ChangePassword', array_keys($_GET))));
+            $req_actions = http_build_query(array_slice($_GET, (int)array_search('ChangePassword', array_keys($_GET))));
 
             $_SESSION['login_redirect'] = [
                 'url' => OSCOM::getLink(null, null, $req_actions, 'SSL')
